requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.29
asyncpg>=0.29.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
